+++ runtime/dice/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 edition = "2018"
-name = "dice-runtime"
+name = "snd-dice-ctl-service"
 version = "0.1.0"
 # For publishing.
 authors = ["Takashi Sakamoto"]
 description = """
-Implementation of service runtime for models based on ASICs of DICE.
+Runtime implementation to control audio and music unit in IEEE 1394 bus which ALSA dice driver
+(snd-dice) supports.
 """
 license = "GPL-3.0-or-later"
 publish = false
